fix: Fix macOS 15+ compilation by upgrading Conan deps and removing workarounds
Upgrade C++ third-party dependencies (protobuf 5.27, grpc 1.67, abseil 20250127,
google-cloud-cpp 2.28, opentelemetry-cpp 1.23, folly 2024.08, fmt 11, glog 0.7,
libcurl 8.10, openssl 3.3, etc.) to versions compatible with macOS 15+.

Key changes:
- Remove Homebrew aws-sdk-cpp workaround; use Conan aws-sdk-cpp on all platforms
- Remove knowhere SVE/hook.cc cmake patches (fixed upstream in knowhere 418f0b6)
- Remove milvus-storage Homebrew AWS SDK shim and Boost isolation hacks
- Fix folly SharedMutex lock holder API (use std::unique_lock/std::shared_lock)
- Fix OpenTelemetry WithActiveSpan API change (nostd::shared_ptr wrapper)
- Fix google-cloud-cpp ComputeEngineCredentials constructor change
- Fix fmt 11 requiring const format methods
- Add static_cast for enum-to-int formatting (fmt 11 strictness)
- Add TARGET_OS_OSX=1 define for Homebrew LLVM on macOS
- Cap MACOSX_DEPLOYMENT_TARGET for Rust cc-rs compatibility
- Update protoc version to 5.27.0 in all generated .pb.go files

+++ internal/core/src/exec/expression/GISFunctionFilterExpr.cpp
@@ -345,7 +345,7 @@ PhyGISFunctionFilterExpr::EvalForDataSegment() {
         default: {
             ThrowInfo(NotImplemented,
                       "internal error: unknown GIS op : {}",
-                      expr_->op_);
+                      static_cast<int>(expr_->op_));
         }
     }
     return res_vec;
@@ -455,7 +455,9 @@ PhyGISFunctionFilterExpr::EvalForIndexSegment() {
                 // Distance-based operation - no prepared version
                 return left.dwithin(query_geometry, expr_->distance_);
             default:
-                ThrowInfo(NotImplemented, "unknown GIS op : {}", expr_->op_);
+                ThrowInfo(NotImplemented,
+                          "unknown GIS op : {}",
+                          static_cast<int>(expr_->op_));
         }
     };
